+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## [unreleased]
 
+- #14:
+
+  - Adds a `provided` dependency on SCI to `pom.xml`, so that cljdoc builds
+    succeed.
+
+  - Upgrades to Clerk version `fad499407d979916d21b33cc7e46e73f7a485e37` for the
+    template and project docs notebook, along with a matching clerk-utils
+    upgrade.
+
+  - Remove `aliases` argument from `mafs.sci/install!`, as these won't work
+    anymore with the new Clerk upgrade.
+
+  - Fixes to docs in template
+
 ## 0.2.0
 
 - #12 adds a `deps-new` template that sets up a basic Clerk project with